Update README.md
Used R shiny to create interactive visualizations on Cryptocurrency. Cryptocurrency is digital currency, the first taste of virtual money, available to anyone, anywhere, anytime. The visualizations we created shows the volatility, which is the dips and highs of the currency. People tend to buy more when there is a dip in the price. The data used for our analysis was free and datasource is Kaggle. We got the data from multiple different kaggle projects. Since we merged more than two datasets, we had to do a little bit of data pre-processing before we started working on the app.